+++ benches/benchmark.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use criterion::{criterion_group, criterion_main, Criterion, Throughput};
 
-use beancount_parser::parse;
+use beancount_parser::BeancountFile;
 
 const SAMPLE: &str = include_str!("../tests/samples/official.beancount");
 
@@ -11,7 +11,7 @@ pub fn run_bench(c: &mut Criterion) {
     group.significance_level(0.01);
     group.throughput(Throughput::Bytes(SAMPLE.len() as u64));
     group.measurement_time(Duration::from_secs(10));
-    group.bench_function("parse", |b| b.iter(|| parse::<f64>(SAMPLE)));
+    group.bench_function("parse", |b| b.iter(|| SAMPLE.parse::<BeancountFile<f64>>()));
     group.finish();
 }