+++ src/event.rs
@@ -6,9 +6,9 @@ use crate::{string, IResult, Span};
 ///
 /// # Example
 /// ```
-/// # use beancount_parser::DirectiveContent;
+/// # use beancount_parser::{BeancountFile, DirectiveContent};
 /// let input = r#"2023-05-31 event "Location" "Switzerland""#;
-/// let beancount = beancount_parser::parse::<f64>(input).unwrap();
+/// let beancount: BeancountFile<f64> = input.parse().unwrap();
 /// let DirectiveContent::Event(ref event) = beancount.directives[0].content else { unreachable!() };
 /// assert_eq!(event.name, "Location");
 /// assert_eq!(event.value, "Switzerland");