+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * `Entry` enum which is one of `Directive`, `Option`, `Include`
 * `parse_iter` which return an iterator over `Result<Entry<D>, Result>`
 * implement `Clone` for `Error`
+* implement `FromStr` for `BeancountFile<D>` where `D: Decimal`
 
 ## [2.0.0-beta.1] - 2023-07-01